[CL-3384] Move notifications to its own file for reuse
